requirements.txt
absl-py==2.3.0
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
dotenv==0.9.9
flatbuffers==25.2.10
gast==0.6.0
google-auth==2.40.3
google-genai==1.23.0
google-pasta==0.2.0
grpcio==1.73.1
h11==0.16.0
h5py==3.14.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
keras==3.10.0
libclang==18.1.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.1.0
numpy==2.1.3
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
requests==2.32.4
rich==14.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tenacity==8.5.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
